Moved OT algorithms from operations classes to separate module named transformOperation.
