+++ tests/document/operation/changeoperation.js
@@ -4,7 +4,6 @@
  */
 
 /* bender-tags: document */
-/* bender-include: ../../_tools/tools.js */
 /* global describe, before, beforeEach, it, expect */
 
 /* bender-include: ../../_tools/tools.js */
@@ -15,36 +14,25 @@ const getIteratorCount = bender.tools.core.getIteratorCount;
 
 const modules = bender.amd.require(
 	'document/document',
-	'document/node',
 	'document/operation/changeoperation',
-	'document/operation/insertoperation',
-	'document/operation/moveoperation',
-	'document/operation/nooperation',
 	'document/position',
 	'document/range',
 	'document/character',
 	'document/attribute',
-	'document/nodelist',
 	'document/text',
 	'ckeditorerror'
 );
 
 describe( 'ChangeOperation', () => {
-	let Document, Node, ChangeOperation, InsertOperation, MoveOperation, NoOperation, Position, Range,
-		Character, Attribute, NodeList, Text, CKEditorError;
+	let Document, ChangeOperation, Position, Range, Character, Attribute, Text, CKEditorError;
 
 	before( () => {
 		Document = modules[ 'document/document' ];
-		Node = modules[ 'document/node' ];
 		ChangeOperation = modules[ 'document/operation/changeoperation' ];
-		InsertOperation = modules[ 'document/operation/insertoperation' ];
-		MoveOperation = modules[ 'document/operation/moveoperation' ];
-		NoOperation = modules[ 'document/operation/nooperation' ];
 		Position = modules[ 'document/position' ];
 		Range = modules[ 'document/range' ];
 		Character = modules[ 'document/character' ];
 		Attribute = modules[ 'document/attribute' ];
-		NodeList = modules[ 'document/nodelist' ];
 		Text = modules[ 'document/text' ];
 		CKEditorError = modules.ckeditorerror;
 	} );
@@ -388,742 +376,6 @@ describe( 'ChangeOperation', () => {
 			let conflicts = op.conflictsAttributesWith( op2 );
 
 			expect( conflicts ).to.be.true;
-		} );
-	} );
-
-	describe( 'getTransformedBy', () => {
-		let nodeA, nodeB, start, end, range, oldAttr, newAttr, op, baseVersion, expected;
-		let expectOperation;
-
-		beforeEach( () => {
-			nodeA = new Node();
-			nodeB = new Node();
-
-			baseVersion = doc.version;
-
-			oldAttr = new Attribute( 'foo', 'abc' );
-			newAttr = new Attribute( 'foo', 'bar' );
-
-			expected = {
-				type: ChangeOperation,
-				oldAttr: oldAttr,
-				newAttr: newAttr,
-				baseVersion: baseVersion + 1
-			};
-
-			expectOperation = bender.tools.operations.expectOperation( Position, Range );
-		} );
-
-		describe( 'on multi-level range', () => {
-			beforeEach( () => {
-				start = new Position( [ 1, 2 ], root );
-				end = new Position( [ 2, 2, 4 ], root );
-
-				range = new Range( start, end );
-
-				op = new ChangeOperation( range, oldAttr, newAttr, baseVersion );
-
-				expected.range = new Range( start.clone(), end.clone() );
-			} );
-
-			describe( 'InsertOperation', () => {
-				it( 'target at different position: no operation update', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 3, 3, 2 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target at offset before: increment offset', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 1, 1 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.offset += 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target at same offset: increment offset', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 1, 2 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.offset += 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target at offset after: no operation update', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 3, 2 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target before node from path: increment index on path', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 0 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.path[ 0 ] += 2;
-					expected.range.end.path[ 0 ] += 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target after node from path: no position change', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 2, 6 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target inside change range: split into two operations', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 1, 3, 1 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.start.path = [ 1, 3, 3 ];
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start = op.range.start;
-					expected.range.end.path = [ 1, 3, 1 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-			} );
-
-			describe( 'ChangeOperation', () => {
-				it( 'attributes not conflicting: no operation update', () => {
-					let transformBy = new ChangeOperation(
-						range.clone(),
-						new Attribute( 'abc', true ),
-						new Attribute( 'abc', false ),
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'is less important: no operation update', () => {
-					let transformBy = new ChangeOperation(
-						new Range( new Position( [ 1, 1, 4 ], root ), new Position( [ 3 ], root ) ),
-						oldAttr,
-						null,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy, true );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range contains original range: convert to NoOperation', () => {
-					let transformBy = new ChangeOperation(
-						new Range( new Position( [ 1, 1, 4 ], root ), new Position( [ 3 ], root ) ),
-						oldAttr,
-						null,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], {
-						type: NoOperation,
-						baseVersion: baseVersion + 1
-					} );
-				} );
-
-				// [ original range   <   ]   range transformed by >
-				it( 'range intersects on left: shrink original range', () => {
-					let transformBy = new ChangeOperation(
-						new Range( new Position( [ 1, 4, 2 ], root ), new Position( [ 3 ], root ) ),
-						oldAttr,
-						null,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.end = new Position( [ 1, 4, 2 ], root );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				// [  range transformed by  <   ]  original range  >
-				it( 'range intersects on right: shrink original range', () => {
-					let transformBy = new ChangeOperation(
-						new Range( new Position( [ 1 ], root ), new Position( [ 2, 1 ], root ) ),
-						oldAttr,
-						null,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start = new Position( [ 2, 1 ], root );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range is inside original range: split original range', () => {
-					let transformBy = new ChangeOperation(
-						new Range( new Position( [ 1, 4, 1 ], root ), new Position( [ 2, 1 ], root ) ),
-						oldAttr,
-						null,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.end.path = [ 1, 4, 1 ];
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.path = [ 2, 1 ];
-					expected.range.end = op.range.end;
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-			} );
-
-			describe( 'MoveOperation', () => {
-				it( 'range and target are different than change range: no operation update', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 1, 1, 2 ], root ),
-						new Position( [ 3, 4 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range offset is before change range start offset: decrement offset', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 1, 0 ], root ),
-						new Position( [ 3, 4, 1 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.offset -= 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target offset is before change range start offset: increment offset', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 3, 4, 1 ], root ),
-						new Position( [ 1, 0 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.offset += 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range is before node from path to change range: decrement index on path', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 0 ], root ),
-						new Position( [ 2, 4, 1 ], root ),
-						1,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.path[ 0 ]--;
-					expected.range.end.path[ 0 ]--;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range is after node from path to change range: no position change', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 3 ], root ),
-						new Position( [ 0, 1 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target before node from path to change range: increment index on path', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 3, 4, 1 ], root ),
-						new Position( [ 1, 0 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.path[ 1 ] += 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target after node from path to change range: no position change', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 3, 4, 1 ], root ),
-						new Position( [ 3 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range intersects on left with change range: split into two operations', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 2, 1 ], root ),
-						new Position( [ 4 ], root ),
-						3,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.end.path = [ 2, 1 ];
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.path = [ 4 ];
-					expected.range.end.path = [ 5, 4 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-
-				it( 'range intersects on right with change range: split into two operation', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 1, 1 ], root ),
-						new Position( [ 0, 0 ], root ),
-						3,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.start.path = [ 1, 1 ];
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.path = [ 0, 1 ];
-					expected.range.end.path = [ 0, 3 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-
-				it( 'range contains change range: update change range', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 1 ], root ),
-						new Position( [ 3, 4, 1 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.path = [ 1, 4, 1, 2 ];
-					expected.range.end.path = [ 1, 4, 2, 2, 4 ];
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range is inside change range: split into two operations', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 1, 4 ], root ),
-						new Position( [ 3, 2 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.path = [ 3, 2 ];
-					expected.range.end.path = [ 3, 4 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-
-				it( 'target inside change range: split into two operations', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 3, 4, 1 ], root ),
-						new Position( [ 1, 4 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.start.path = [ 1, 6 ];
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start = op.range.start;
-					expected.range.end.path = [ 1, 4 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-
-				it( 'range intersects change range and target inside change range: split into three operations', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 1, 1 ], root ),
-						new Position( [ 2 ], root ),
-						3,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 3 );
-
-					expected.range.start.path = [ 5 ];
-					expected.range.end.path = [ 5, 2, 4 ];
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.path = [ 1, 1 ];
-					expected.range.end.path = [ 2 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-
-					expected.range.start.path = [ 3 ];
-					expected.range.end.path = [ 5 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 2 ], expected );
-				} );
-			} );
-
-			describe( 'NoOperation', () => {
-				it( 'no operation update', () => {
-					let transformBy = new NoOperation( baseVersion );
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-			} );
-		} );
-
-		// Some extra cases for a ChangeOperation that operates on single tree level range.
-		// This means that the change range start and end differs only on offset value.
-		describe( 'on single-level range', () => {
-			beforeEach( () => {
-				start = new Position( [ 0, 2, 1 ], root );
-				end = new Position( [ 0, 2, 4 ], root );
-
-				range = new Range( start, end );
-
-				op = new ChangeOperation( range, oldAttr, newAttr, baseVersion );
-
-				expected.range = new Range( start.clone(), end.clone() );
-			} );
-
-			describe( 'InsertOperation', () => {
-				it( 'target at offset before: increment offset', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 0, 2, 0 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.offset += 2;
-					expected.range.end.offset += 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target at same offset: increment offset', () => {
-					let transformBy = new InsertOperation(
-						new Position( [ 0, 2, 1 ], root ),
-						[ nodeA, nodeB ],
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.offset += 2;
-					expected.range.end.offset += 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-			} );
-
-			describe( 'MoveOperation', () => {
-				it( 'range offset is before change range start offset: decrement offset', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 0, 2, 0 ], root ),
-						new Position( [ 2, 4, 1 ], root ),
-						1,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.offset--;
-					expected.range.end.offset--;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target offset is before change range start offset: increment offset', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 2, 4, 1 ], root ),
-						new Position( [ 0, 2, 0 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.offset += 2;
-					expected.range.end.offset += 2;
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range intersects on left with change range: split into two operations', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 0, 2, 2 ], root ),
-						new Position( [ 2, 4, 1 ], root ),
-						4,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.end.offset -= 2;
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.path = [ 2, 4, 1 ];
-					expected.range.end.path = [ 2, 4, 3 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-
-				it( 'range intersects on right with change range: split into two operation', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 0, 2, 0 ], root ),
-						new Position( [ 2, 4, 1 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.start.offset -= 1;
-					expected.range.end.offset -= 2;
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.path = [ 2, 4, 2 ];
-					expected.range.end.path = [ 2, 4, 3 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-
-				it( 'range contains change range: update change range', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 0, 1 ], root ),
-						new Position( [ 2, 4, 1 ], root ),
-						3,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start.path = [ 2, 4, 2, 1 ];
-					expected.range.end.path = [ 2, 4, 2, 4 ];
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'range is inside change range: split into two operations', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 0, 2, 2 ], root ),
-						new Position( [ 2, 4, 1 ], root ),
-						1,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.end.offset--;
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.path = [ 2, 4, 1 ];
-					expected.range.end.path = [ 2, 4, 2 ];
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-
-				it( 'range is same as change range: update change range', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 0, 2, 1 ], root ),
-						new Position( [ 2, 4, 1 ], root ),
-						3,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expected.range.start = new Position( [ 2, 4, 1 ], root );
-					expected.range.end = new Position( [ 2, 4, 4 ], root );
-
-					expectOperation( transOp[ 0 ], expected );
-				} );
-
-				it( 'target inside change range: split into two operations', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 2, 4, 1 ], root ),
-						new Position( [ 0, 2, 2 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 2 );
-
-					expected.range.start.offset = 4;
-					expected.range.end.offset = 6;
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.offset = op.range.start.offset;
-					expected.range.end.offset = 2;
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-				} );
-
-				it( 'range intersects change range and target inside change range: split into three operations', () => {
-					let transformBy = new MoveOperation(
-						new Position( [ 0, 2, 0 ], root ),
-						new Position( [ 0, 2, 3 ], root ),
-						2,
-						baseVersion
-					);
-
-					let transOp = op.getTransformedBy( transformBy );
-
-					expect( transOp ).to.be.instanceof( Array );
-					expect( transOp.length ).to.equal( 3 );
-
-					expected.range.start.offset = 3;
-					expected.range.end.offset = 4;
-
-					expectOperation( transOp[ 0 ], expected );
-
-					expected.range.start.offset = 0;
-					expected.range.end.offset = 1;
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 1 ], expected );
-
-					expected.range.start.offset = 2;
-					expected.range.end.offset = 3;
-					expected.baseVersion++;
-
-					expectOperation( transOp[ 2 ], expected );
-				} );
-			} );
 		} );
 	} );
 } );
