+++ tests/document/operation/nooperation.js
@@ -9,27 +9,15 @@
 
 const modules = bender.amd.require(
 	'document/document',
-	'document/position',
-	'document/range',
-	'document/attribute',
-	'document/operation/nooperation',
-	'document/operation/insertoperation',
-	'document/operation/changeoperation',
-	'document/operation/moveoperation'
+	'document/operation/nooperation'
 );
 
 describe( 'NoOperation', () => {
-	let Document, Position, Range, Attribute, NoOperation, InsertOperation, ChangeOperation, MoveOperation;
+	let Document, NoOperation;
 
 	before( function() {
 		Document = modules[ 'document/document' ];
-		Position = modules[ 'document/position' ];
-		Range = modules[ 'document/range' ];
-		Attribute = modules[ 'document/attribute' ];
 		NoOperation = modules[ 'document/operation/nooperation' ];
-		InsertOperation = modules[ 'document/operation/insertoperation' ];
-		ChangeOperation = modules[ 'document/operation/changeoperation' ];
-		MoveOperation = modules[ 'document/operation/moveoperation' ];
 	} );
 
 	let noop, doc, root;
@@ -39,13 +27,6 @@ describe( 'NoOperation', () => {
 		doc = new Document();
 		root = doc.createRoot( 'root' );
 	} );
-
-	function expectNoTransformation( noop, transformBy ) {
-		const transOp = noop.getTransformedBy( transformBy );
-
-		expect( transOp ).to.be.instanceof( NoOperation );
-		expect( transOp.baseVersion ).to.equal( 0 );
-	}
 
 	it( 'should not throw an error when applied', () => {
 		expect( () => doc.applyOperation( noop ) ).to.not.throw( Error );
@@ -63,38 +44,5 @@ describe( 'NoOperation', () => {
 
 		expect( clone ).to.be.an.instanceof( NoOperation );
 		expect( clone.baseVersion ).to.equal( 0 );
-	} );
-
-	it( 'should not change when transformed by InsertOperation', () => {
-		const transformBy = new InsertOperation( new Position( [ 0 ], root ), 'abc', 0 );
-
-		expectNoTransformation( noop, transformBy );
-	} );
-
-	it( 'should not change when transformed by ChangeOperation', () => {
-		const transformBy = new ChangeOperation(
-			new Range( new Position( [ 0 ], root ), new Position( [ 1 ], root ) ),
-			null,
-			new Attribute( 'foo', 'bar' ),
-			0
-		);
-
-		expectNoTransformation( noop, transformBy );
-	} );
-
-	it( 'should not change when transformed by MoveOperation', () => {
-		const transformBy = new MoveOperation(
-			new Position( [ 0 ], root ),
-			new Position( [ 1 ], root ),
-			4,
-			0
-		);
-
-		expectNoTransformation( noop, transformBy );
-	} );
-
-	it( 'should not change when transformed by NoOperation', () => {
-		const transformBy = new NoOperation( 0 );
-		expectNoTransformation( noop, transformBy );
 	} );
 } );
